Initial implementation of fakenet.
Fakenet is intended to be a simple fake (in-memory) implementation of
POSIX network functions. This way, we can test toxcore without actually
running on a real network, which greatly improves test times and
flakiness. We can then run one test with a real network and all others
with the fake network.

Alternatively (or additionally), this layer can forward calls to the
real system functions, modifying them to create some separation between
test runs.